Clean up and deprecate lib/telemetry
