+++ apps/studio/components/ui/PageTelemetry.tsx
@@ -1,3 +1,4 @@
+import { Sha256 } from '@aws-crypto/sha256-browser'
 import * as Sentry from '@sentry/nextjs'
 import { useRouter } from 'next/router'
 import { PropsWithChildren, useEffect } from 'react'
@@ -5,8 +6,16 @@ import { PropsWithChildren, useEffect } from 'react'
 import { useUser } from 'common'
 import { useSendPageMutation } from 'data/telemetry/send-page-mutation'
 import { IS_PLATFORM, LOCAL_STORAGE_KEYS } from 'lib/constants'
-import { getAnonId } from 'lib/telemetry'
 import { useAppStateSnapshot } from 'state/app-state'
+
+const getAnonId = async (id: string) => {
+  const hash = new Sha256()
+  hash.update(id)
+  const u8Array = await hash.digest()
+  const binString = Array.from(u8Array, (byte) => String.fromCodePoint(byte)).join('')
+  const b64encoded = btoa(binString)
+  return b64encoded
+}
 
 const PageTelemetry = ({ children }: PropsWithChildren<{}>) => {
   const user = useUser()
